LayoutTests/fast/events/onunload-clears-onbeforeunload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: before unload
CONSOLE MESSAGE: unload
If you didn't hit an assert you PASS.
